Correctly setup wallet default onion and local forwarding ports
This PR removes a port collision if the user has an explicitly set
`tor_forward_address` set in their config. This option now only applies
to the base node. The wallet is always configured to use an OS-assign
forwarding port and a pre-defined wallet onion port of 18101 (the same
as the Aurora wallet).