Unblinded outout semantics
This commit encourages the use of UnblindedOutput everywhere one needs
to perform, operations using UTXO data. Fields are made private to
improve forward compatibility.

Eventually, the only way to create a TransactionOutput will be via
UnblindedOutput::as_transaction_output(), but the constructor is left in
for now.

The sign! macro is also introduced to make future changes to kernel
signature semantics easier to manage. If additional firleds are needed
in future to create a kernel signature, but the defaults work in most
cases, then this macro will be very handy in reducing the number of
changes that have to be made throughout the code.