Unify key semantics
The "spending key" was being referenced from
UnblindedOutput.sepnding_key() is some spots and .blinding_factor() in
others.

Only the spending key part of the blinding factor is secret, so let's just store the spending key in
the db under the spending key column. It reduces the total change, and
also reduces any refactoring from future changes to the commitment /
blinding factor definition.