Integrate LibWallet into the `tari_console_wallet` application
This PR integrates the tari_wallet backend into the tari_console_wallet application. The application needed to reused a lot of the utility functions from `tari_base_node` so it moves those utility functions into a new crate called `tari_app_utilities`. 

The wallet will use the same config location as `tari_base_node` and the same identity and data files. It will create a separate log config and log to a separate folder within ~/.tari.

This PR is still in progress and the following is what I think still needs to be done to flesh out all the functionality present in the current UI:

- Add way to specific wallet base node in config (Currently just using the first Seed Node)
- Add way to edit base node pub key and address in console UI
- Read Balance from wallet
- Read Contacts from wallet
- Update above on relevant wallet events by subscribing to the wallet Event Streams.
- Read Chain tip from Base Node in wallet (this is a whole thing) and plumb into UI with updates via wallet events
- Add calling Send transaction in wallet
- Add contacts from UI
- Read current connections from Wallet for Network Tab
- Optimize the cacheing and display of the Complete Transaction list which lags when it is very long. Currently it caches the full completed transactions but only needs a few fields which should help a lot.
- Determine if running this app with the old BN will create issues with a single database file in terms of crashes and corruption (Perhaps the WAL journal mode is good enough to allow this operation in Sqlite?)