Add `is_synced` flag to BaseNodeServiceResponse & update UTXO validator
This PR adds a Boolean flag to all BaseNodeServiceResponse’s so that the node querying can know if the queried node is synced to the chain tip. The wallet is then updated to abort a UTXO validation operation if this flag is false. This is because if a node is not synced it could incorrectly declare UTXO invalid that it is not yet aware of.

To do this the Base Node Service needed to know whether the Base Node State machine had synced at least once and was currently in the Listening state. To allow the Base Node Service to speak to the Base Node State Machine the state machine had to be refactored into a Service with a Handle that could be registered in the Service Stack.

Additional the base node was updated to not process new blocks or transactions until it had synced at least once. This determination was done using the flag made available via the new BaseNodeStateMachineHandle.

Tests are provided that test:
1. That the flag is set in the base node state machine once it has synced
2. That the Output Manager Service UTXO validation is aborted if the `is_synced` flag is false.

A test was not written to check that the `is_synced` flag in BaseNodeServiceResponse is set based on the info provided by the BaseNodeStateMachineHandle because the infrastructure to write this test does not exist and will be a huge project to write. This was however tested by compiling and running on a live node.