Refactor the bootstrapped flag into StatusInfo
Currently the bootstrapped state is only available in the ListeningInfo status but we need to be able to access it in all the states. 
This PR refactors the StatusInfo enum to a struct with the old StatusInfo enum becoming StateInfo. StatusInfo now contains the old StateInfo as a field and bootstrapped as a field. This way the full state machine status is available via the StatusInfo watch channel.