Allow client application to control when transaction broadcast protocols are restarted
Previously the wallet restarted its transaction broadcast and negotiation protocols as soon as a base node pubkey was set. This caused problems in the base node where on a restart the mempool was empty but the chain had yet to sync so validly mined transactions were cancelled because their chain monitoring protocol started before the blocks were synced.

This PR separates the functions to restart the transaction protocols from the set_base_node_public_key function in the transaction service. This allows the transaction negotiation protocols to be started ASAP as they don’t need a base node and then the client application can choose when to start the broadcast protocols.

The base node application was updated to wait for the local node to be synced before starting the transaction broadcast protocols. The FFI library was updated to start these protocol right after the base node public key was supplied to retain the previous behaviour in the mobile wallets.