Refactor basenode/wallet/p2p initialization
The main thrust of ths PR is to allow services assembled using the StackBuilder
to add protocols to comms before comms is spawned. These protocols will
typically be RPC services.

There existed a _chicken-and-egg_ problem between initializing the
`StackBuilder` services and RPC services which may depend on those services,
which in turn may depend on comms services.

The solution in this PR is to initialize comms itself using the stack
builder (`tari_p2p::P2pInitializer`) allowing objects to be registered
on the `StackBuilder` before comms has spawned. A `SharedCommsContext`
is provided to services that wish to register comms protocols and some
comms components which are available during early intialization (e.g.
PeerManager).

- Created `P2pInitializer` and added it to the base node and wallet
  `StackBuilder`
- Refactored generics out of the `CommsBuilder` and `UnspawnedCommsNode`
- Greatly improved ergonomics of `StackBuilder`
- Added `spawn_when_ready` convinience function to `ServiceInitializer` context
  that allows a service to wait for all handles to become ready before
  starting.
- Comms no longer provides `ShutdownSignals` but has to have a shutdown
  signal provided to it. This allows for a single shutdown signal per
  app.
- Cleanup of `tari_base_node::builder` module
- Removed redundant `Shutdown` trigger. A single shutdown trigger is
  used throughout
- Refactored tests and examples as necessary