Improvements for tie breaking comms, connection manager cleanup
Deduplicate the connection state management within comms.
Previously, `ConnectionManager` and `ConnectivityManager` both held `PeerConnection` handles.
This was because the `Connectivity` concept came after the connection
manager and was not refactored at the time. Now the connection state is
contained only within `ConnectivityManager` and `ConnectionManager` is
soley responsible for establishing outbound connections and handling
inbound connections.

Connections closed by tie breaking are now delayed by 2s and the
`PeerConnectionWillClose` emitted to allow components to react to a
connection that will close. The prime use case is the messaging protocol
which can finish sending remaining messages and cut off further messages
from being lost before the connection is closed.

This should increase messaging reliability and possibly a fix for #2295