Small refactor of `fetch_headers`
`fetch_headers` takes a `start` and `end_inclusive` numbers instead of
`Vec<u64>`. In all usages of `fetch_headers` the array was sequential.
This limits the API to only what is needed and so allows more optimal
implementations to exist (e.g. `SELECT * from headers where height >= x
AND height <= y`). However, the main reason for this is not wanting to
include this refactor in a future block sync implementation PR.