Add automated stress test to base node wallet
This change adds the ability to perform an automated stress test that uses a
command file as input. The command file accommodates one optional `coin-split`
command and one or more `make-it-rain` commands. The number and amount of
test UTXOs will be assessed, and if found to be inadequate to perform the
entire test, the `coin-split` will be attempted if requested. In the event of
a `coin-split`, the wallet will be monitored for up to 30 minutes to determine
if a sufficient number of UTXOs became available, and when that is achieved,
the stress test will start. The `make-it-rain` commands can be scheduled to
start at a specific time, thus the test can be started in advance.