Update TUI Transaction listing and Contact management
This PR firstly cleans up the TUI implementation a bit further from the initial refactor now that we are working with real data from the backend rather than dummy data. This involves a little bit or reorganisation. The main contributions of this PR are an updated StatefulList that can provide a windowed view into the list data and contact management.

The update to the StatefulList implementation was required because the naive implementation put forward by tui-rs will parse and format the entire list for every draw operation. This was a problem on a wallet with thousands of transactions which made the UI very laggy. This PR adds functionality to the Stateful list which will produce a windowed view into the list data based on what items are selected and the provided height of the list widget that will be rendered so only the minimum required items are formatted for rendering. To avoid rewriting the whole List widget from tui-rs this is achieved by outputting the windowed slice of the data and a generated ListState object that the native tui-rs render method expects.

The second part of this PR is adding the implementation to allow for managing Contacts in the Send/Receive tab. This involved completing the UI elements for listing contacts and adding, editing or deleting contacts. The initial UI logic for selecting a contact to send to, adding an amount and hitting Send Transaction is added though only the validation is done and not the actual sending yet. Errors are also displayed using an Error popup dialog.

Tested:
Tests are provided for the updates to the StatefulList and the windowed view.