Add a client key-value store to the wallet and wallet_ffi
The mobile team that uses the WalletFFI library requested a feature to be able to store arbitrary Key-Values in the Wallet’s database and make use of the Wallet’s encryption for these values.

To enable this the functionality to set, get and clear client key-value pairs was added to the Wallet database as a new table. If an encryption cipher is set for the wallet the Values will be encrypted and decrypted by the database backend.

Three FFI interface functions were added to the FFI library to allow a client to access this functionality.

How it is tested:
Tests are provided for the functions in the database backend and in the FFI library.