Revert change to fetch_mmr_nodes from #2357
The change to `fetch_mmr_nodes` introduced in #2357 caused the
`request_and_response_fetch_mmr_node_and_count` test to timeout and
fail. This commit reverts that changes, and makes the previous `debug`
log into `warn` for greater visibility.

This really depends on desired behaviour - the test indicates that an
empty set should be returned if a peer requests "out of bounds" MMR
node(s) - instead of an error result. Either this specific change
introduced in PR #2357 needs to be reverted to maintain the test
invariant, or the test itself needs to be changed.