Increase lagging broadcast channels' capacities
During adverse conditions it was found that the state events and transaction
service broadcast channel capacities were inadequate to handle the load.
Subscribers lagged up to 92 state events (capacity of 10) and 490 transaction
events (capacity of 200). As a rule of thumb the capacities were increased
sufficiently to accommodate 5 to 10 times more events than the peak amount
they lagged by.

```
2020-10-21 19:25:57.722798300 [c::bn::state_machine_service::states::listening]
    DEBUG Metadata event subscriber lagged by 92 item(s)
```
```
2020-10-21 22:44:13.858514800 [base_node::app::utils] WARN  Error reading from
    event broadcast channel Lagged(490)
```

**Edit**: Changes to the transaction service broadcast channel reverted.