Implement wallet base node service
Adds a base node service to the wallet, to start tracking a connected
base node's state. This will be required for the wallet to know the
base node status, chain height, and when last the information was
updated.

The service has a configurable refresh interval to call
`GetChainMetadata` on the base node.