Implement pending transaction cancellation in the TUI app
This PR implements the ability to cancel a pending transaction from the TUI app. If you have pending transactions you can select a transaction and hit C to cancel it. The PR also adds in Confirmation dialogs to the TUI app and implements from the tx cancellation, transaction sending and contact deletion.

This PR also fixes a bug in the WindowedListState logic where if an item in a list was deleted while the last element was selected it broke the List State.