[feature] Ban peers that flood messages
Ban a peer if the number of messages received over a timespan
(`DhtConfig::flood_ban_timespan`) exceeds a threshold
(`DhtConfig::flood_ban_max_msg_count`).

The current defaults are an average of 10mps over 100 seconds.

**How it works**

A layer is added to the incoming pipeline that queues up metrics
(producer) for the metric collector (consumer).
The metric collector is a consumer that reads from a channel and updates
it's in-memory state. The collector is lock-free.

The Dht connectivity actor periodically (2 minutes) queries the
collector for peer's exceeding the configured rate and bans them.