Fix for network discovery state machine spin loop
If the node is unable to obtain many peers it continues to aggresively
ask for peers. On a small localhost network this can cause high CPU
usage as it repeatedly queries for peers in a state loop. This fix adds
checks in the `Ready` state that allow a transition to `Idle` if: there
are no sync peers, there are no sync peers other than the ones used last
time, or there have been 10 rounds in a row where the node has not been
able to obtain "enough" peers to enter "on connect" mode.