Add config option to disable wallet in Base Node
This PR adds the ability to specify the “enable_wallet” boolean config option, which when set too false will not startup the Wallet instance embedded in the Base node.  All commands that used the embedded wallet instance were updated to take this into account.

This is done so that it is possible to run our Seed nodes without the Wallet instance which was cluttering the logs.