Revert change to fetch blocks utxos txos (#2357)
- The change to fetch_utxos, fetch_txos and fetch_blocks introduced in #2357
caused the those calls to timeout and fail in the client, as nothing was
returned if matching data could not be found, contrary to the original design.
This commit reverts that change, and renames the affected `NodeCommsRequest`
functions so their intent can be clear.
- The change to `NodeCommsRequest::fetch_mmr_nodes` was reverted in #2364; this
PR also renames that function aptly.