Simplify automated stress test
Simplified the automated stress test so that it will also automatically
perform any required coin split beforehand. If not enough large valued
UTXOs are available for the coin split, it will create what is possible
and let the user know afterwards what the status is, without performing
the stress test.