Fix wallet conversion error for a valid tx status
Fixes a subtle error where a valid `TransactionStatus` of 5 was
mistakenly reported as a conversion error. This caused the wallet to
fail on startup if coinbase transactions were in teh database.

The `TransactionStatus::try_from` method was missing the value.

Also added some additional info to conversion error, in case it appears
again it will be easier to narrow down.