Add optional fee_per_gram field to console wallet Send Tx form
This PR adds an optional editable fee_per_gram field to the send transaction form in the console wallet.

This field is populated by the default value specified by a constant in LibWallet and is not included in the flow of fields edited while sending a transaction as editing this field should be very rare. If the user wants to edit the field they can use the hotkey and after this transaction is sent the field will be repopulated with the default value again.

This functionality was not able to be tested as the network is currently experiencing some issues that make sending transactions difficult.