Implement entry and persistence of custom base node in console wallet
This PR adds the functionality to allow a user to specify a custom base node public key and address in the Network tab of the console wallet. If the public key and address can be parsed into a Comms Public Key and valid Net address this bae node is set to be used by all the wallet services and is saved into the wallet’s client key-value store.

On startup if there is a valid public key and address stored in the wallets client key-value store this is used as the base node for the wallet to connect to. If these values don’t exist or cannot be parsed into a valid public key or net address the wallet will fallback to the base node specified in the config.

A function is also added to clear the custom base node which, after a confirmation dialog, will delete the custom data from the database and revert to the config specified base node.