Provide initial sync status to merge mining proxy
Initial sync status is now provided via gRPC to the merge mining proxy; this will
enable more efficient use of computing resources when the initial blockchain sync
of the linked base node has not been achieved yet. In this implementation the merge
mining proxy will wait for initial sync to be achieved before instructing XMRig to
start mining, as otherwise it cannot provide a valid Tari block template to mine on.