Improve UX for Windows first runtime use
This change enhances the user experience when id and/or log config
files are missing. Prior to this change, log config files were not
created automatically due to not using the `--init` flag when
starting the Tari executables. Additionally, if only the base node
wallet id file was missnig, it was not detected. These are now fixed.