[wallet] Fix bug where cancelled output is spent
Fixes #2443

During merge mining, sometimes a number of cancelled outputs are added to
the wallet before the mined coinbase utxo.

When spending this valid utxo, the _cancelled_ output was marked as
encumbered and then spent, instead of the valid utxo.

This led to the bug where after "spending" a utxo, it would still exist in the db
as unspent. It would then be selected for transactions, but reported as spent and be unspendable.

Fundamentally the issue was that `OutputSql::find_by_commitment` was not
excluding cancelled outputs.