Add wallet build step to catch regressions
The libwallet build requires that the `tari_wallet` crate can be built
on independently of the base node. This commit adds
`cargo build -p tari_wallet` to the circleci build step to ensure that
future commits don't break this requirement.

As a part of this, the `waiting_requests` module was
extracted to tari_common_types since it is used in both the wallet and
base node (and was previously duplicated in both).