Don’t spawn multiple monitoring tasks for the reused coinbase tx
Currently if multiple coinbase transactions are requested for the same block height the wallet will reuse the prior transaction but a whole new monitoring protocol task will be spawned for that coinbase tx even though one is already running. 

This PR adds TxId to be stored with the protocols monitoring base node responses and if one exists for a newly requested coinbase that is reusing a transaction a new monitoring protocol is not started.