Save tor identity file in console wallet
If a tor identity file does not exist a new one is generated, however it
is up to the calling code to persist the file in some way. This PR adds
some code to persist the tor identity file, as well as the comms
identity file containing the public address of the node.