Fix cucumber tests for Windows use
Windows port numbers higher than 49999 found to be a problem,
specifically the static assignment of port 50071 for the merge
mining proxy server. This fix allows selection of free port
numbers in the range 19000 to 25000 for all instances of GRPC
servers, using existing `getFreePort` function. This also
allows using the merge mining proxy in consecutive tests as the
port is not freed up after a single test for multiple test
usage.