Remove generated proto files from tari_core
Cleans up the code a bit in tari_core::proto by removing the `generated`
module. Instead of including the generated files directly, protos are
generated to the OUT_DIR and included using the `include!` macro. This
is the intended way that generated files are included, however
this would break intellisense in IDEs.

IDE and rust analyser support for OUT_DIR has greatly improved.
see:
https://blog.jetbrains.com/clion/2019/12/recent-updates-in-rust/
https://github.com/rust-analyzer/rust-analyzer/pull/3549