Remove "add one to port" hack in console wallet
The "add one to port" hack was used in the base node wallet to ensure
that port numbers don't conflict in the _base node wallet_ due to
deficiencies in the current configuration framework.

This PR removes this hack from the console wallet code (only effects TCP
transport) as it creates confusion when implementing cucumber tests.