Update fee vs amount being sent validation
Currently when the `SenderTransactionInitializer` validates the components of the transaction it is building it will check that if the fee is more than the amount being sent + the change. 

This PR updates this behaviour to check if the fee being sent is more than just the amounts being send, ignoring the change. This prevents the case where a large UTXO that produces a lot of change will still allow a transaction to be sent where the outbound amount is less than the fee.