Fix tests breaking due to Fee vs Sent amount validation
In a previous PR which changed the transaction builder validation to check that the fee of a transaction is not more than the amount being sent (as opposed to the total amount). This broke a lot of tests due to the helper functions to make transaction building the transaction without using the Amounts part of the transaction builder. Another feature that uses the transaction builder without a recipient is the Coinsplit.

In order to fix this, this PR only applies the Fee vs Sent amount validation if there is an amount being sent.