Implement WalletNotify for console wallet TUI mode
Adds a Notifier module and hooks into the console wallet TUI event
monitor to trigger a notify script when transactions are sent, received,
cancelled, mined with x confirmations, and completely mined and
confirmed.

By launching the TUI with a path given `--notify` command line argument, or the
`notify` key in the `[wallet]` config file, the given path will be
executed with available args from the transaction. See
`notify_example.sh` file for the args.