Replace P2P comms with RPC for Coinbase Monitoring
This PR reworks the coinbase monitoring protocol to be in line with the
latest wallet transaction validation protocol (PR#2614) and to make use
of RPC comms rather than P2P comms. The two P2P request/response interactions
to the base node, get chain metadata and fetch UTXOs, have been replaced by
a single RPC call to query if the coinbase kernel, identified by the signature,
exists. This call also returns the best block height that is needed by the
coinbase monitoring protocol to decide if the coinbase transaction should be
cancelled or not. Also updated integration tests.